fix(cli): address pr review feedback
